refactor(website): scope generic Twenty API lib to partners
Addresses review feedback that 'twenty-api' / TWENTY_API_* were too
generic for what is really a partners-only client:
- rename lib/twenty-api -> lib/partners-api
- rename env vars TWENTY_API_URL/KEY -> TWENTY_PARTNERS_API_URL/KEY
- rename twentyFetch -> partnersApiFetch and stop exporting it from the
  barrel, so the lib's public surface is just getPartners + types
