+++ packages/twenty-website/.env.example
@@ -36,6 +36,7 @@ ENTERPRISE_JWT_PUBLIC_KEY=
 # Optional: short-lived validity token length in days (default 30)
 # ENTERPRISE_VALIDITY_TOKEN_DURATION_DAYS=
 
-# Twenty workspace the marketing site reads data from (server-side only)
-TWENTY_API_URL=
-TWENTY_API_KEY=
+# Twenty workspace the partners marketplace reads partner data from
+# (server-side only) via the /s/partners REST endpoint.
+TWENTY_PARTNERS_API_URL=
+TWENTY_PARTNERS_API_KEY=
