+++ packages/twenty-website/src/app/[locale]/partners-marketplace/__tests__/filter-partners.test.ts
@@ -1,4 +1,4 @@
-import type { MarketplacePartner } from '@/lib/twenty-api';
+import type { MarketplacePartner } from '@/lib/partners-api';
 
 import {
   EMPTY_CRITERIA,
