+++ packages/twenty-website/src/app/[locale]/partners-marketplace/components/PartnerCard.tsx
@@ -7,7 +7,7 @@ import { theme } from '@/theme';
 import { styled } from '@linaria/react';
 import type { CSSProperties } from 'react';
 
-import type { MarketplacePartner } from '@/lib/twenty-api';
+import type { MarketplacePartner } from '@/lib/partners-api';
 import { PartnerAvatar } from './PartnerAvatar';
 import { PartnerChipRow } from './PartnerChipRow';
 import {
