+++ packages/twenty-website/src/app/[locale]/partners-marketplace/filter-url-helpers.ts
@@ -2,7 +2,7 @@ import type {
   DeploymentExpertise,
   ServedGeo,
   SpokenLanguage,
-} from '@/lib/twenty-api';
+} from '@/lib/partners-api';
 
 import type { FilterCriteria } from './filter-partners';
