+++ packages/twenty-website/src/lib/partners-api/index.ts
@@ -1,0 +1,7 @@
+export { getPartners } from '@/lib/partners-api/get-partners';
+export type {
+  DeploymentExpertise,
+  MarketplacePartner,
+  ServedGeo,
+  SpokenLanguage,
+} from '@/lib/partners-api/partner-types';
