+++ packages/twenty-website/src/lib/partners-api/partner-types.ts
