--- packages/twenty-website/src/lib/twenty-api/index.ts
@@ -1,8 +1,0 @@
-export { twentyFetch } from '@/lib/twenty-api/client';
-export { getPartners } from '@/lib/twenty-api/get-partners';
-export type {
-  DeploymentExpertise,
-  MarketplacePartner,
-  ServedGeo,
-  SpokenLanguage,
-} from '@/lib/twenty-api/partner-types';
